Log-In and Basal CSS Changes
What I did today was add some extra functionalities that are going to exist in the Log-In Page (e.g. remember me, forgot password) with some minor tweaks to the CSS Ashley made. I also added a reference to the index.html in the G-PLAN title in the top-left so its more ergonomic to move back to the home page. Lastly, I created the majorsMinors.css that will be referred by both the major and minor html branches. It's currently just a copy-pasta of the indexStyle.css until changes are made to the major selection plan idea. Also tweaked a typo in the original index.html that wasn't referencing majors.html correctly.