Userbase & Sign-Up Expanded Functionality
Fleshed out the sign-up functionality with new methods and new JavaScript files containing and managing the userbase and its contained user objects saved locally. Still needs some extra work, but the ground for it is there and should be easy applying saved user data to everything that is needed. Ask me if you have any questions regarding it.